requirements.txt
vanna[milvus,ollama,postgres]
fastapi
uvicorn[standard]
python-dotenv
pandas>=2.0.0
plotly>=5.0.0
kaleido
python-multipart
pydantic
jinja2
pyngrok
openpyxl
watchfiles
langdetect>=1.0.9
networkx>=3.0
langchain>=0.1.0
sentence-transformers>=2.2.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
langchain_community
googletrans